+++ src/main/java/com/cobber/fta/plugins/LogicalTypeMonthAbbr.java
@@ -74,7 +74,7 @@ public class LogicalTypeMonthAbbr extends LogicalTypeFinite {
 	}
 
 	@Override
-	public String isValidSet(final String dataStreamName, final long matchCount, final long realSamples, final FactsTypeBased facts, final Map<String, Long> cardinality, final Map<String, Long> outliers, final Shapes shapes, AnalysisConfig analysisConfig) {
+	public String isValidSet(final String dataStreamName, final long matchCount, final long realSamples, String currentRegExp, final FactsTypeBased facts, final Map<String, Long> cardinality, final Map<String, Long> outliers, final Shapes shapes, AnalysisConfig analysisConfig) {
 		if (outliers.size() > 1)
 			return LocaleInfo.getShortMonthsRegExp(locale);