+++ src/main/java/com/cobber/fta/plugins/LogicalTypePhoneNumber.java
@@ -102,7 +102,7 @@ public class LogicalTypePhoneNumber extends LogicalTypeInfinite  {
 	}
 
 	@Override
-	public String isValidSet(final String dataStreamName, final long matchCount, final long realSamples, final FactsTypeBased facts, final Map<String, Long> cardinality, final Map<String, Long> outliers, final Shapes shapes, AnalysisConfig analysisConfig) {
+	public String isValidSet(final String dataStreamName, final long matchCount, final long realSamples, String currentRegExp, final FactsTypeBased facts, final Map<String, Long> cardinality, final Map<String, Long> outliers, final Shapes shapes, AnalysisConfig analysisConfig) {
 		if (getHeaderConfidence(dataStreamName) == 0 && cardinality.size() <= 20 || getConfidence(matchCount, realSamples, dataStreamName) < getThreshold()/100.0)
 			return REGEXP;
 		return null;