+++ src/test/java/com/cobber/fta/PluginBirthDate.java
@@ -78,7 +78,7 @@ public class PluginBirthDate extends LogicalTypeInfinite {
 
 	@Override
 	public String isValidSet(final String dataStreamName, final long matchCount, final long realSamples,
-			final FactsTypeBased facts, final Map<String, Long> cardinality, final Map<String, Long> outliers, Shapes shapes, AnalysisConfig analysisConfig) {
+			String currentRegExp, final FactsTypeBased facts, final Map<String, Long> cardinality, final Map<String, Long> outliers, Shapes shapes, AnalysisConfig analysisConfig) {
 		return (double)matchCount/realSamples >= getThreshold()/100.0 ? null : ".+";
 	}
 }